More MSVC fixes plus cleanup warnings a bit

+++ sleepyhead/SleepLib/loader_plugins/resmed_loader.cpp
@@ -149,7 +149,7 @@ qint16 EDFParser::Read16()
 
 QString EDFParser::Read(unsigned n)
 {
-    if ((pos + n) > filesize) {
+    if ((pos + long(n)) > filesize) {
         return "";
     }
